+++ QLExtension/PreviewViewController.swift
@@ -124,50 +124,7 @@ class PreviewViewController: NSViewController, QLPreviewingController {
             
             let extrajs: String
             if settings.unsafeHTMLOption && settings.inlineImageExtension {
-                extrajs = """
-<script type="text/javascript">
-/**
- * Generate an unique id.
- * @returns {string}
- */
-function uid() {
-    let a = new Uint32Array(3);
-    window.crypto.getRandomValues(a);
-    return (performance.now().toString(36)+Array.from(a).map(A => A.toString(36)).join("")).replace(/\\./g,"");
-}
-
-/**
- * Callback event handler for error when loading an image.
- */
-function handleImageError() {
-    if (this.src.indexOf("file://") === 0) {
-        // Process only local images.
-        if (!this.id) {
-            this.id = 'I' + uid(); // Assign a uid.
-        }
-        window.webkit.messageHandlers.imageExtensionHandler.postMessage({src: this.src, id: this.id});
-    }
-}
-
-/**
- * Replace the src image with the embedded data.
- */
-function replaceImageSrc(result) {
-    const img = document.getElementById(result.id);
-    if (img) {
-        img.src = result.data;
-        return true;
-    } else {
-        return false;
-    }
-}
-
-// Register the onerror handler.
-for (let image of document.images) {
-    image.onerror = handleImageError
-}
-</script>
-"""
+                extrajs = "<script type=\"text/javascript\">" + (settings.getBundleContents(forResource: "inlineimages", ofType: "js") ?? "") + "</script>\n";
             } else {
                 extrajs = ""
             }